Add snapshot tests for ChannelHeader previews
Extract each preview body into an internal helper so the
ChannelHeaderTest can drive snapshots from the same composables
that back the @Preview annotations, mirroring the pattern used
in ChannelMediaAttachmentsScreen.

Moves the test next to its source under
ui.messages.header, adds a thread-mode preview, and adds a
few-members preview to keep member-count coverage that the old
test had with channelWithFewMembers.

Drops the user-typing preview/test: typing is rendered as a
MessageListTypingIndicatorItemContent in the message list since
PR #6206, so the header's typingUsers parameter no longer
produces a visible change.
