+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/ui/messages/header/ChannelHeaderTest.kt
@@ -1,0 +1,71 @@
+/*
+ * Copyright (c) 2014-2026 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.chat.android.compose.ui.messages.header
+
+import app.cash.paparazzi.DeviceConfig
+import app.cash.paparazzi.Paparazzi
+import io.getstream.chat.android.compose.ui.PaparazziComposeTest
+import org.junit.Rule
+import org.junit.Test
+
+internal class ChannelHeaderTest : PaparazziComposeTest {
+
+    @get:Rule
+    override val paparazzi = Paparazzi(deviceConfig = DeviceConfig.PIXEL_2)
+
+    @Test
+    fun connected() {
+        snapshotWithDarkMode {
+            ChannelHeaderConnected()
+        }
+    }
+
+    @Test
+    fun connecting() {
+        snapshotWithDarkMode {
+            ChannelHeaderConnecting()
+        }
+    }
+
+    @Test
+    fun offline() {
+        snapshotWithDarkMode {
+            ChannelHeaderOffline()
+        }
+    }
+
+    @Test
+    fun `few members`() {
+        snapshotWithDarkMode {
+            ChannelHeaderFewMembers()
+        }
+    }
+
+    @Test
+    fun `many members`() {
+        snapshotWithDarkMode {
+            ChannelHeaderManyMembers()
+        }
+    }
+
+    @Test
+    fun `thread mode`() {
+        snapshotWithDarkMode {
+            ChannelHeaderThreadMode()
+        }
+    }
+}
