+++ stream-chat-android-compose/src/test/snapshots/images/io.getstream.chat.android.compose.ui.messages.header_ChannelHeaderTest_few_members.png
